+++ lib/octicons.rb
@@ -1,9 +1,13 @@
 require "octicons/version"
 require "octicons/octicon"
 require "nokogiri"
+require "json"
 
 module Octicons
   OCTICONS_SVG_PATH = File.join(File.dirname(__FILE__), "svg/*.svg")
+
+  KEYWORDS = JSON.parse(File.read(File.join(File.dirname(__FILE__), "keywords.json")))
+  CODEPOINTS = JSON.parse(File.read(File.join(File.dirname(__FILE__), "codepoints.json")))
 
   OCTICON_SYMBOLS = {}
   Dir[OCTICONS_SVG_PATH].each do |svg_path|