+++ lib/octicons/octicon.rb
@@ -13,7 +13,7 @@ module Octicons
         @height = octicon[:height]
 
         # create html_options from options, except for a few
-        @html_options = @options.reject { |d| [:size].include? d }
+        @html_options = @options
         @html_options.merge!({
           :class   => classes,
           :viewBox => viewbox,
@@ -26,8 +26,29 @@ module Octicons
       end
     end
 
+    # Returns an string representing a <svg> tag
     def to_svg
       "<svg #{html_attributes}>#{@path}</svg>"
+    end
+
+    # Returns an array of keywords similar to the icon
+    def keywords
+      Octicons::KEYWORDS[@symbol]["keywords"]
+    end
+
+    # Returns the decimal codepoint of the character
+    def decimal
+      Octicons::CODEPOINTS[@symbol]
+    end
+
+    # Returns the hexidecimal version of the character
+    def hexadecimal
+      decimal.to_s(16)
+    end
+
+    # Returns the unicode character
+    def character
+      [decimal].pack("U")
     end
 
     private
@@ -67,18 +88,8 @@ module Octicons
         :height => @height
       }
 
-      # When size is "large"
-      if @options[:size] == "large"
-        size[:width] = 2 * @width
-        size[:height] = 2 * @height
-
-      # When size is an integer
-      elsif @options[:size].is_a?(Integer) || !!(@options[:size] =~ /\A[0-9]+\z/)
-        size[:width]  = calculate_width(@options[:size])
-        size[:height] = @options[:size]
-
-        # Specific size
-      elsif !@options[:width].nil? || !@options[:height].nil?
+      # Specific size
+      unless @options[:width].nil? && @options[:height].nil?
         size[:width]  = @options[:width].nil?  ? calculate_width(@options[:height]) : @options[:width]
         size[:height] = @options[:height].nil? ? calculate_height(@options[:width]) : @options[:height]
       end