+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # HEAD
 
 - Adding: attr_reader for `:path, :html_options, :width, :height`
+- Adding: `keywords.json` and `codepoints.json` methods to the octicon object. Making them available via the api
+- Deprecating: the use of `:size` for "large" or a number, instead use `:height`
 - Deprecate: the use of `:symbol`, instead the first argument is the symbol
 
 # 1.1.0